+++ Library/Aphid/UI.js
@@ -24,6 +24,7 @@ Aphid.UI = {};
 
 // Views ---------------------------------------------------------------------
 
+//= require "UI/ImageView"
 //= require "UI/PopoverView"
 //= require "UI/LoadingIndicator"
 //= require "UI/AlertView"